Hide internals (#79)
