+++ requests_auth/_errors.py
