+++ tests/success_ntlm/requests_ntlm.py
@@ -4,6 +4,6 @@ class HttpNtlmAuth:
         self.password = password
 
     def __call__(self, r):
-        r.headers[
-            "Authorization"
-        ] = f"HttpNtlmAuth fake {self.username} / {self.password}"
+        r.headers["Authorization"] = (
+            f"HttpNtlmAuth fake {self.username} / {self.password}"
+        )
