+++ tests/test_oauth2_implicit_id_token_okta.py
@@ -1,9 +1,10 @@
 import requests_auth
+import requests_auth._oauth2.implicit
 
 
 def test_corresponding_oauth2_implicit_flow_id_token_instance(monkeypatch):
     monkeypatch.setattr(
-        requests_auth.authentication.uuid,
+        requests_auth._oauth2.implicit.uuid,
         "uuid4",
         lambda *args: "27ddfeed4e-854b-4361-8e7a-eab371c9bc91",
     )
