+++ tests/test_oauth2_resource_owner_password_okta.py
@@ -4,8 +4,8 @@ import pytest
 import requests
 
 import requests_auth
+from requests_auth.testing import token_cache  # noqa: F401
 from tests.auth_helper import get_header, get_request
-from requests_auth.testing import token_cache
 
 
 def test_oauth2_password_credentials_flow_uses_provided_session(
@@ -97,7 +97,7 @@ def test_oauth2_password_credentials_flow_token_is_expired_after_30_seconds_by_d
     token_cache._add_token(
         key="bdc39831ac59c0f65d36761e9b65656ae76223f2284c393a6e93fe4e09a2c0002e2638bbe02db2cc62928a2357be5e2e93b9fa4ac68729f4d28da180caae912a",
         token="2YotnFZFEjr1zCsicMWpAA",
-        expiry=requests_auth.oauth2_tokens._to_expiry(expires_in=29),
+        expiry=requests_auth._oauth2.tokens._to_expiry(expires_in=29),
     )
     # Meaning a new one will be requested
     responses.add(
@@ -138,7 +138,7 @@ def test_oauth2_password_credentials_flow_token_custom_expiry(
     token_cache._add_token(
         key="bdc39831ac59c0f65d36761e9b65656ae76223f2284c393a6e93fe4e09a2c0002e2638bbe02db2cc62928a2357be5e2e93b9fa4ac68729f4d28da180caae912a",
         token="2YotnFZFEjr1zCsicMWpAA",
-        expiry=requests_auth.oauth2_tokens._to_expiry(expires_in=29),
+        expiry=requests_auth._oauth2.tokens._to_expiry(expires_in=29),
     )
     assert (
         get_header(responses, auth).get("Authorization")
